src/backend/core/serializers.py
from rest_framework import serializers
from .models import Activity, Calendar, ChosenActivity
from rest_framework import serializers
from django.contrib.auth.models import User
from rest_framework import serializers
from rest_framework.validators import UniqueTogetherValidator


class UserSerializer(serializers.ModelSerializer):

    def create(self, validated_data):
        user = User.objects.create_user(**validated_data)
        return user

    class Meta:
        model = User
        fields = (
            'id',
            'username',
            'first_name',
            'last_name',
            'email',
            'password',
        )
        validators = [
            UniqueTogetherValidator(
                queryset=User.objects.all(),
                fields=['username', 'email']
            )
        ]

class ActivitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Activity
        fields = ['id', 'location', 'category', 'description', 'source_link']

class CalendarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Calendar
        fields = ['id', 'user', 'name']

class ChosenActivitySerializer(serializers.ModelSerializer):
    class Meta:
        model = ChosenActivity
        fields = ['id', 'activity', 'calendar', 'user', 'start_date', 'end_date']
