src/backend/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
Django==5.1.2
djangorestframework==3.15.2
matplotlib==3.9.2
protobuf==4.25.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-decouple==3.8
setuptools==75.1.0
sqlparse==0.5.1
wheel==0.44.0
